Create test cases
Do take note of that traffic light has shading variety for yellow sign. A portion of the nations have that as golden or silver shading. Some may even discover the shading distinctive all together. So everything relies upon which standard that is being pursued for the traffic signal machine and nation traffic rules. You might need to modify the traffic governs according to the standards applied in your state or nation. That would make the beneath experiments increasingly precise and compelling. 
•	Red light– stop and wait before the zebra crossing line.
•	Green light – go.
•	Yellow light– stop, unless – 1) you’ve already crossed the stop line after green light is off. 2) you’re so close to it that  might cause an accident. 3) You’re in middle of the road and should procced to next side rather than going back causing traffic problem.
The list of test cases for traffic signal:
Check if the traffic lights are on during the normal traffic time frame according to rules. 
Check if the traffic lights clock functions according to anticipated explicit sign. 
Check if the traffic lights are shown in consecutive request. 
Check if the traffic light for stop signal is appeared with red light. 
Check if the traffic light for hold up except if signal is appeared with yellow or golden shading. 
Check if the traffic light for go is appeared with green. 
Check if the traffic lights are each in turn. 
Check if the traffic lights during off hours show flickering yellow or golden shading showing off for rules. 
Check the time delay between the lights. 
Check the arrangement of numerous sign posts are in a state of harmony in go across streets.